refactor: apply cr suggestions
