+++ packages/react-ui-kit/src/Form/RangeInput.tsx
@@ -21,62 +21,14 @@
 import {CSSObject, jsx} from '@emotion/react';
 import React, {FC, forwardRef} from 'react';
 import {TextProps} from '../Text';
-import {COLOR_V2} from '../Identity';
-import {manySelectors} from '../util';
 import {Theme} from '../Layout';
 import InputLabel from './InputLabel';
-
-const rangeInputWrapperStyles: CSSObject = {
-  position: 'relative',
-  marginTop: '19px',
-};
-
-const thumbSelectors = ['&::-webkit-slider-thumb', '&::-moz-range-thumb', '&::-ms-thumb'];
-const sliderSelectors = ['&::-webkit-slider-runnable-track', '&::-moz-range-track', '&::-ms-track'];
-
-const getImageCropZoomInputStyles = (theme: Theme, backgroundSize: `${number}% ${number}%`): CSSObject => ({
-  display: 'block',
-  '-webkit-appearance': 'none',
-  width: '100%',
-  height: '8px',
-  background: COLOR_V2.GRAY_60,
-  borderRadius: '4px',
-  backgroundImage: `linear-gradient(${theme.general.primaryColor}, ${theme.general.primaryColor})`,
-  backgroundSize: backgroundSize || '0% 100%',
-  backgroundRepeat: 'no-repeat',
-
-  ...manySelectors(thumbSelectors, {
-    '-webkit-appearance': 'none',
-    height: '18px',
-    width: '18px',
-    borderRadius: '50%',
-    background: COLOR_V2.GRAY_80,
-    cursor: 'pointer',
-    border: 'none',
-    boxShadow: 'none',
-  }),
-
-  ...manySelectors(sliderSelectors, {
-    '-webkit-appearance': 'none',
-    boxShadow: 'none',
-    border: 'none',
-    background: 'transparent',
-  }),
-});
-
-enum ValueLabelPosition {
-  LEFT = 'left',
-  RIGHT = 'right',
-}
-
-const getValueLabelStyles = (position: ValueLabelPosition): CSSObject => ({
-  pointerEvents: 'none',
-  bottom: '100%',
-  fontSize: '16px',
-  fontWeight: 300,
-  position: 'absolute',
-  [position]: '4px',
-});
+import {
+  getImageCropZoomInputStyles,
+  getValueLabelStyles,
+  rangeInputWrapperStyles,
+  ValueLabelPosition,
+} from './RangeInput.styles';
 
 export interface RangeInputProps<T = HTMLInputElement> extends TextProps<T> {
   label?: string;
@@ -99,7 +51,11 @@ export const RangeInput: FC<RangeInputProps> = forwardRef<HTMLInputElement, Rang
     wrapperCSS,
     ...inputProps
   }) => {
-    const backgroundSize = `${((Number(value) - Number(min)) * 100) / (Number(max) - Number(min))}% 100%` as const;
+    const minNum = Number(min);
+    const maxNum = Number(max);
+    const valueNum = Number(value);
+
+    const backgroundSize = `${((valueNum - minNum) * 100) / (maxNum - minNum)}% 100%` as const;
 
     return (
       <div css={wrapperCSS}>
